requirements.txt
langgraph 
langchain-openai 
langchain-community
python-docx
python-dotenv
autogen
pypdf
